requirements.txt
pandas
apache-airflow==2.8.1
psycopg2-binary==2.9.9
scikit-learn==1.4.0 
streamlit
